Implement zthrottle inspired throttling
This will implement a zthrottle inspired throttling that will space out
requests of a burst with N seconds in between, ignoring all but the
first and the last request of the burst.
This will make sure the last push of a burst is never swallowed and
always reaches the device while still throttling the burst.